Refactoring various things in the code base (#29)
* Reduces sizes and padding

* Removed unneeded bool in GLTexture

* Created method for loading textures

* Removed unneeded vars

* Reduced size of GLBufferLayout

* Added manual destuction of texture

+++ Editor/include/Input/KeyCodes.h
@@ -2,7 +2,7 @@
 
 #include <cstdint>
 
-using KeyCode = uint32_t;
+using KeyCode = uint16_t;
 
 namespace Key {
 	enum : KeyCode {
